Implement end to end encryption, migrate to typescript

+++ backend/src/models/File.ts
@@ -41,12 +41,12 @@ export async function getMetadataFromPath(path: string) {
 }
 
 /**
- * Download a specified file as a buffer
+ * Get a specified file as a buffer
  * @param id the user's id
  * @param name the file name
  * @returns the file buffer
  */
-export async function download(id: number, name: string) {
+export async function get(id: number, name: string) {
   const path = id + "/" + name;
   const [buffer] = await bucket.file(path).download();
 
@@ -54,11 +54,11 @@ export async function download(id: number, name: string) {
 }
 
 /**
- * Download a specified shared file as a buffer
+ * Get a specified shared file as a buffer
  * @param fileId the shared file's id
  * @returns the file buffer
  */
-export async function downloadShared(fileId: string) {
+export async function getShared(fileId: string) {
   const path = "shared/" + fileId;
   const [buffer] = await bucket.file(path).download();
 
@@ -69,9 +69,10 @@ export async function downloadShared(fileId: string) {
  * Upload a file to Google Cloud Storage
  * @param id the user's id
  * @param file the file to upload
+ * @param iv the initialization vector for the file if encrypted
  * @returns the upload response from Google Cloud Storage
  */
-export async function upload(id: number, file: UploadedFile) {
+export async function upload(id: number, file: UploadedFile, iv?: Buffer) {
   const filePath = file.tempFilePath;
   const fileId = uniqid();
   const destFileName = id + "/" + file.name;
@@ -81,14 +82,19 @@ export async function upload(id: number, file: UploadedFile) {
     metadata: {
       metadata: {
         id: fileId,
+        iv: iv?.toJSON().data.toString(),
       },
     },
   });
+
+  if (!iv) {
+    await share(id, file.name);
+  }
   return res;
 }
 
 /**
- * Delete a file from Google Cloud Storage. If the file was shared,
+ * Delete a file from Google Cloud Storage. If the file is public,
  * it will also be deleted from the shared folder.
  * @param id the user's id
  * @param name the file name
@@ -96,11 +102,15 @@ export async function upload(id: number, file: UploadedFile) {
  */
 export async function trash(id: number, name: string) {
   const destFileName = id + "/" + name;
-  const fileId = (await getMetadata(id, name)).metadata.id;
+  const { metadata } = await getMetadata(id, name);
+  const fileId = metadata.id;
   const res = await bucket.file(destFileName).delete();
 
-  const destShared = "shared/" + fileId;
-  await bucket.file(destShared).delete();
+  const isPublic = !metadata.iv;
+  if (isPublic) {
+    const destShared = "shared/" + fileId;
+    await bucket.file(destShared).delete();
+  }
 
   return res;
 }
